# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: QuoteCraft API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/users:
    get:
      operationId: listAdminUsers
      tags: [admin]
      summary: Owner-only combined view of all registered users
      description: >
        Syncs the local user_profiles table from Clerk, then returns every
        registered user with email, sign-up method, registration/closure
        times, marketing consent, and their business names. Requires the
        signed-in user's email to be listed in the ADMIN_EMAILS environment
        variable.
      responses:
        "200":
          description: All registered users with consent and businesses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminUser"
        "403":
          description: Not an admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /billing/checkout:
    post:
      operationId: createBillingCheckout
      tags: [billing]
      summary: Start a Stripe Checkout session (Pro subscription or credit pack)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BillingCheckoutInput"
      responses:
        "200":
          description: Checkout session URL, or resumed subscription
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BillingCheckoutResult"
  /billing/confirm:
    post:
      operationId: confirmBillingCheckout
      tags: [billing]
      summary: Confirm a completed Checkout session (idempotent fulfillment)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BillingConfirmInput"
      responses:
        "200":
          description: Fulfillment result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BillingConfirmResult"
  /billing/cancel:
    post:
      operationId: cancelBillingSubscription
      tags: [billing]
      summary: Cancel the Pro subscription at the end of the billing period
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BillingCancelInput"
      responses:
        "200":
          description: Cancellation scheduled
          content:
            application/json:
              schema:
                type: object
                required: [result]
                properties:
                  result:
                    type: string
                  currentPeriodEnd:
                    type: [string, "null"]
  /usage/pdf-download:
    post:
      operationId: consumePdfDownload
      tags: [billing]
      summary: Authorize and record one PDF download (consumes quota or credit)
      responses:
        "200":
          description: Download allowed
          content:
            application/json:
              schema:
                type: object
                required: [allowed, source]
                properties:
                  allowed:
                    type: boolean
                  source:
                    type: string
                    enum: [pro, credit, free]
                  creditsRemaining:
                    type: integer
        "402":
          description: Free tier limit reached
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LimitReachedError"
components:
  schemas:
...
    BillingCheckoutInput:
      type: object
      required: [type]
      properties:
        type:
          type: string
          enum: [subscription, credits]
        credits:
          type: integer
          description: Pack size (10, 20, 50 or 100) when type is credits
    BillingCheckoutResult:
      type: object
      properties:
        url:
          type: [string, "null"]
        resumed:
          type: boolean
    BillingConfirmInput:
      type: object
      required: [sessionId]
      properties:
        sessionId:
          type: string
    BillingCancelInput:
      type: object
      properties:
        reason:
          type: string
    BillingConfirmResult:
      type: object
      required: [result]
      properties:
        result:
          type: string
          enum: [subscription_active, credits_added]
        creditsAdded:
          type: integer
        credits:
          type: integer
    LimitReachedError:
      type: object
      required: [error, code, action]
      properties:
        error:
          type: string
        code:
          type: string
          enum: [LIMIT_REACHED]
        action:
          type: string
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    AdminUser:
      type: object
      properties:
        userId:
          type: string
        email:
          type: string
        signupMethod:
          type: ["string", "null"]
        registeredAt:
          type: ["string", "null"]
          format: date-time
        closedAt:
          type: ["string", "null"]
          format: date-time
        marketingConsent:
          type: boolean
        businesses:
          type: array
          items:
            type: string
      required:
        - userId
        - email
        - signupMethod
        - registeredAt
        - closedAt
        - marketingConsent
        - businesses
...